refactor + configuration + fixed
