+++ MoodProject.App/wwwroot/appsettings.Development.json
@@ -1,6 +1,6 @@
 ﻿{
   "Cache": {
-    "Version": "DEV3"
+    "Version": "dev"
   },
   "Api": {
     "BaseUrl": "https://localhost:7073/api/"
@@ -9,6 +9,8 @@
     "HubUrl": "https://localhost:7073/notifications"
   },
   "Quizz": {
-    "IgnoreMinDaysToGenerateQuizz": true
+    "IgnoreMinDaysToGenerateQuizz": true,
+    "MinDaysBeforeNewQuizz": 1,
+    "MaxDaysBeforeNewQuizz": 7
   }
 }
