+++ MoodProject.App/wwwroot/appsettings.Production.json
@@ -9,6 +9,8 @@
     "HubUrl": "https://moodproject.azurewebsites.net/notifications"
   },
   "Quizz": {
-    "IgnoreMinDaysToGenerateQuizz": true
+    "IgnoreMinDaysToGenerateQuizz": true,
+    "MinDaysBeforeNewQuizz": 1,
+    "MaxDaysBeforeNewQuizz": 7
   }
 }
