+++ MoodProject.Core/Ports/In/IQuizzGenerator.cs
@@ -1,0 +1,8 @@
+﻿using MoodProject.Core.Models;
+
+namespace MoodProject.Core.Ports.In;
+
+public interface IQuizzGenerator
+{
+    public Task<OperationResult<IEnumerable<QuizzQuestion>>> Generate(string userId);
+}
