+++ MoodProject.Services/Configuration/QuizzConfiguration.cs
@@ -3,4 +3,6 @@
 public class QuizzConfiguration
 {
     public bool IgnoreMinDaysToGenerateQuizz { get; set; }
+    public int MinDaysBeforeNewQuizz { get; set; }
+    public int MaxDaysBeforeNewQuizz { get; set; }
 }
